New icon made: aside.png to show / hide the sidebar ...
The other files in images/icon have *not* been changed but
appear here because of the Makefile they have a new timestamp.

The icon aside.png appears in the menu.

It is intended to show / hide the sidebar.

There has been the discussion whether such a feature should not be
system-wide or only for certain themes.

The feature is *theoretically* theme-independent, however
it has no effect if not in the according theme it is in the
nav.tpl.

So de facto it is only in the theme vier.

And: successfully also in theme mobil-vier although no change there.

Seemingly here the inheritance is functioning properly.

Th aside.png icon at present time has not yet any functionality because
it points to the friendica module that shows the information.

***

mobil-vier: not beautiful - avatar and search textfield move down -
but at least usable? Testing wanted!

At present time no pull request; will try to add the functionality to the icon.

David