Still not really functional and not beautiful - icon reachable
Moved the nav menu a little bit down.

The toggler for mobile / not mobile appears; had been hidden!

... maybe somebody long before me did this :-)

Some elements removed "float right" command.

Still no functionality for the icon.

The icon was not reachable when the text of the banner is long
and is reachable now.

Screenshot follows as comment.