Deleting and creating leftsidebar database entry - EXPERIMENTAL!
Found that there is a general default.php file and every theme may
have its own or not.

So modified file in vier and vier-mobil; original /view/default.php unchanged.

Planned, but not implemented, is - when user clicks toggle sidebar icon -

that the pconfig entry "system leftsidebar show" is deleted and the
result is: from then on the left sidebar is not shown.

When user again clicks toggle sidebar button, the pconfig entry
system leftsidebar show is inserted, and the result is
that the sidebar is shown.

The next assumption is that when user leaves the theme vier or vier-mobil,

the programm will "listen" to the "general" default.php file which
always shows the sidebar.

I believe that in one week I will be ready with it.