Continued work on theme vier-mobil - need some help
Text of mobile toggler replaced by new icon.

Is not yet functional.

The icon must point to a new module that will be similar to toggle_mobile,
or even simpler. Module is not yet ready.

Change to style.css: nav menu and sidebar now scroll.

Testing on computer by changing line 566 in boot.php to ... = true;

Testing on Nokia N97 mobile phone gives blank page; however,
after restart of the *Web server* it appears.

Cause: Unknown.

Is not yet beautiful.

Shortage of knowledge on CSS.

Need some help here!

On the N97 the sidebar scrolls vertically but not with the keys
but only with finger swiping; reason unknown.

Bug in the browser of the N97??

Desirable: make the mobile toggler and sidebar toggler icons
with bright color instead of black, and
transparent background color instead of white.

Possible integrate the two icons into the icon font?

Have no knowledge about that.