Hiding - on demand - sidebar on vier and vier-mobil
The idea at present time is:

By z-index: 3 for section, that is one higher than for aside,
one will not see the sidebar, if the position of widget
is left: 0px, or not given.

Module toggle_leftaside sets an entry for to show sidebar, or deletes it.

When sidebar - width is 215px - is to be shown, the position of
section is set to left: 215

Not everything ready.