Update compose.tpl
The changes start the editor in a larger editor window. In addition, the authorisation settings for the article are hidden for the time being. These can be shown via a sandwitch menu if required.

The Composer Editor is another editor in Friendica, which is particularly suitable if longer texts are to be entered (e.g. at the local workstation). The special feature is that the editor is loaded in a separate window, thus avoiding distractions from the timeline.